Improvements.

+++ MoneySendEscrow.sol
@@ -4,8 +4,8 @@ pragma solidity ^0.8.7;
 
 contract MoneySendEscrow {
     // Define variables.
-    uint public balanceReceived;
-    uint public lockedUntil;
+    uint256 public balanceReceived;
+    uint256 public lockedUntil;
 
     // Contract address receive money.
     function receiveMoney() public payable {
@@ -14,13 +14,13 @@ contract MoneySendEscrow {
     }
 
     // View received balance.
-    function getBalance() public view returns(uint) {
+    function getBalance() public view returns (uint256) {
         return address(this).balance;
     }
 
     // Withdraw received balance.
     function withdrawMoney() public {
-        if(lockedUntil < block.timestamp) {
+        if (lockedUntil < block.timestamp) {
             address payable to = payable(msg.sender);
             to.transfer(getBalance());
         }
@@ -28,7 +28,7 @@ contract MoneySendEscrow {
 
     // Withdraw funds to specific address.
     function withdrawMoneyTo(address payable _to) public {
-        if(lockedUntil < block.timestamp) {
+        if (lockedUntil < block.timestamp) {
             _to.transfer(getBalance());
         }
     }
