+++ WhitelistMapping.sol
@@ -4,11 +4,11 @@ pragma solidity ^0.8.7;
 
 contract WhitelistMapping {
     // Declare mappings.
-    mapping(uint => bool) public myMapping;
+    mapping(uint256 => bool) public myMapping;
     mapping(address => bool) public myAddressMapping;
-    mapping (uint => mapping(uint => bool)) uintUintBoolMapping;
+    mapping(uint256 => mapping(uint256 => bool)) uintUintBoolMapping;
 
-    function setValue(uint _index) public {
+    function setValue(uint256 _index) public {
         myMapping[_index] = true;
     }
 
@@ -16,11 +16,19 @@ contract WhitelistMapping {
         myAddressMapping[msg.sender] = true;
     }
 
-    function setUintUintBoolMapping(uint _index1, uint _index2, bool _value) public {
+    function setUintUintBoolMapping(
+        uint256 _index1,
+        uint256 _index2,
+        bool _value
+    ) public {
         uintUintBoolMapping[_index1][_index2] = _value;
     }
 
-    function getUintUintBoolMapping(uint _index1, uint _index2) public view returns (bool) {
+    function getUintUintBoolMapping(uint256 _index1, uint256 _index2)
+        public
+        view
+        returns (bool)
+    {
         return uintUintBoolMapping[_index1][_index2];
     }
 }
